load data item history

+++ app/src/main/java/com/klcn/xuant/transporter/mvp/history/ListHistoryAdapter.java
@@ -62,7 +62,7 @@ public class ListHistoryAdapter extends BaseAdapter {
         holder.mTxtPlaceLocation.setText(tripInfo.getPickup());
         holder.mTxtPlaceDistination.setText(tripInfo.getDropoff());
         if (tripInfo.getRating() != null)
-            holder.mRating.setRating(Integer.parseInt(tripInfo.getRating()));
+            holder.mRating.setRating(Float.parseFloat(tripInfo.getRating()));
         else
             holder.mRating.setRating(0);
